+++ PanicSystem/Controller.cs
@@ -188,6 +188,5 @@ namespace PanicSystem
             }
             return -1;
         }
-
     }
 }
